20240812_2

+++ printf/ft_check.c
@@ -6,7 +6,7 @@
 /*   By: oyayoi <oyayoi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 17:40:34 by oyayoi            #+#    #+#             */
-/*   Updated: 2024/08/12 14:36:41 by oyayoi           ###   ########.fr       */
+/*   Updated: 2024/08/12 14:47:33 by oyayoi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,13 +46,13 @@ int	ft_divide(int str, va_list args)
 	else if (str == 'p')
 		count += ft_putptr((unsigned long long)va_arg(args, unsigned long long));
 	else if (str == 'd' || str == 'i')
-		count += ft_putnbr(va_arg(args, int));
+		count += ft_putnbr((int)va_arg(args, int));
 	else if (str == 'u')
-		count += ft_putunbr(va_arg(args, unsigned int));
+		count += ft_putunbr((unsigned int)va_arg(args, unsigned int));
 	else if (str == 'x')
-		count += ft_puthex(va_arg(args, unsigned int), 0);
+		count += ft_puthex((unsigned int)va_arg(args, unsigned int), 0);
 	else if (str == 'X')
-		count += ft_puthex(va_arg(args, unsigned int), 1);
+		count += ft_puthex((unsigned int)va_arg(args, unsigned int), 1);
 	else if (str == '%')
 		count += ft_putchar('%');
 	return (count);
